requirements.txt
boto3==1.35.7
python-dotenv==1.0.1